refactor: migrate UpdateAttributeTypesService
ref #2318
