+++ visualization/app/codeCharta/state/effects/updateAttributeTypes/attributeTypes.merger.spec.ts
@@ -1,5 +1,5 @@
 import { getMergedAttributeTypes } from "./attributeTypes.merger"
-import { AttributeTypes, AttributeTypeValue } from "../../../../codeCharta.model"
+import { AttributeTypes, AttributeTypeValue } from "../../../codeCharta.model"
 
 describe("AttributeTypesMerger", () => {
 	describe("getMergedAttributeTypes", () => {
