+++ visualization/app/codeCharta/state/state.module.ts
@@ -5,7 +5,6 @@ import { MapSizeService } from "./store/treeMap/mapSize/mapSize.service"
 import { ScalingService } from "./store/appSettings/scaling/scaling.service"
 import { MarkedPackagesService } from "./store/fileSettings/markedPackages/markedPackages.service"
 import { EdgesService } from "./store/fileSettings/edges/edges.service"
-import { AttributeTypesService } from "./store/fileSettings/attributeTypes/attributeTypes.service"
 import { FocusedNodePathService } from "./store/dynamicSettings/focusedNodePath/focusedNodePath.service"
 import { BlacklistService } from "./store/fileSettings/blacklist/blacklist.service"
 import { InjectorService } from "./injector.service"
@@ -24,7 +23,6 @@ angular
 	.service(camelCase(ScalingService.name), ScalingService)
 	.service(camelCase(MarkedPackagesService.name), MarkedPackagesService)
 	.service(camelCase(EdgesService.name), EdgesService)
-	.service(camelCase(AttributeTypesService.name), AttributeTypesService)
 	.service(camelCase(FocusedNodePathService.name), FocusedNodePathService)
 	.service(camelCase(BlacklistService.name), BlacklistService)
 	.service(camelCase(InjectorService.name), InjectorService)
